+++ split_tests/help_text.py
@@ -1,0 +1,9 @@
+from django.utils.translation import gettext_lazy as _
+
+
+COHORT = {
+    "weight": _(
+        "Enter any positive integer.\n\nFor example, if you want two cohorts with a 75%/25% split, you could enter 75 for"
+        " the first cohort and 25 for the second, or 3 for the first cohort and 1 for the second."
+    )
+}